Split to one class per file
RN needs it so
