+++ lib/queueStatus.ts
@@ -1,0 +1,11 @@
+import { EventEmitter } from 'events'
+
+export class QueueStatus extends EventEmitter {
+  public emitEmptyQueue() {
+    this.emit('EMPTY')
+  }
+
+  public emitIdleQueue() {
+    this.emit('IDLE')
+  }
+}
